+++ firebase-functions/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+* [feature] Streaming callable functions are now supported.
 * [fixed] Fixed an issue that prevented the App Check token from being handled correctly in case of error.